Refactor (WIP)
