+++ src/js/components/app-header.js
@@ -1,3 +1,8 @@
+/**
+ * Filename: 'app-header.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * App header component
+ */
 import React from 'react'
 import SearchArea from './app-search'
 import ListButton from './app-list-button'
@@ -11,14 +16,7 @@ export default () => {
         margin: '10px',
         marginTop: '20px'
     }
-
-    let buttonStyles = {
-        width: 'auto',
-        height: 'auto',
-        marginTop: '18px',
-    }
     return (
-
         <div className="row" style={ {borderBottom: '2px solid #ccc'} }>
             <div className="col-sm-4 text-left">
                 <img src='http://dejanstojanovic.net/media/31519/honey.ico' style={ styles } />
@@ -27,16 +25,7 @@ export default () => {
             <div className="col-sm-7">
                 <SearchArea />
             </div>
-            <div className="col-sm-1">
-                <button 
-                    className="btn btn-info" 
-                    type="button" text="Show List" 
-                    style={buttonStyles} 
-                    onClick="">
-                    <i className="glyphicon glyphicon-user" aria-hidden="true" /> My List
-                </button>
-
-            </div>
+            <ListButton />
         </div>
     )
 }
