+++ src/js/components/app-list-item.js
@@ -1,3 +1,8 @@
+/**
+ * Filename: 'app-list-item.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * App list item component
+ */
 import React from 'react'
 
 export default {
