+++ src/js/components/app-results-item-detail.js
@@ -1,3 +1,8 @@
+/**
+ * Filename: 'app-results-item-detail.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * App results item detail component shows info when item is clicked
+ */
 import React from 'react'
 import AppActions from '../actions/app-actions'
 
@@ -12,7 +17,7 @@ export default ( props ) => {
 	let index = props.index
     return (
         <div className="result" style={ styles }>
-        	Detail...
+        Detail...
         </div>
     )
 }
