+++ src/js/components/app-template.js
@@ -1,3 +1,9 @@
+/**
+ * Filename: 'app-template.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * App template component
+ */
+
 import React from 'react'
 import Header from './app-header'
 import Footer from './app-footer'
