+++ src/js/modules/fetch-data-promise.js
@@ -1,3 +1,9 @@
+/**
+ * Filename: 'fetch-data-promise.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * Returns a promise to get data from server using jQuery.
+ */
+
 import jQuery from 'jquery'
 
 export default ( url ) => {
