+++ src/js/modules/unix-timestamp-formatter.js
@@ -1,3 +1,9 @@
+/**
+ * Filename: 'unix-timestamp-formatter.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * Uses moment npm package to return pretty dates from unix timestamps
+ */
+
 import moment from 'moment'
 
 const FORMAT = "ddd MMM Do, YYYY hh:mm A"
