+++ src/js/static/SourceURL.js
@@ -1,3 +1,9 @@
+/**
+ * Filename: 'SourceURL.js'
+ * Author: JMW <rabbitfighter@cryptolab.net>
+ * Exports the source for getting data from server
+ */
+
 const PROTOCOL = "http://"
 const HOST = "54.200.61.193/"
 const FILE = "clientInfoData.json"
